# perf: optimization of the hot paths