--- a/television/matcher/config.rs
+++ b/television/matcher/config.rs
@@ -8,41 +8,35 @@
 /// matches, and no optimization for matching paths as well as using the
 /// default number of threads (which corresponds to the number of available logical
 /// cores on the current machine).
-#[derive(Copy, Clone, Debug)]
-#[allow(clippy::struct_excessive_bools)]
+#[derive(Debug, Clone)]
 pub struct Config {
-    /// The number of threads to use for the fuzzy matcher.
-    pub n_threads: Option<usize>,
-    /// Whether to ignore case when matching.
-    pub ignore_case: bool,
     /// Whether to prefer prefix matches.
     pub prefer_prefix: bool,
     /// Whether to optimize for matching paths.
-    pub match_paths: bool,
+    pub optimize_for_paths: bool,
+    /// The number of threads to use for matching.
+    pub n_threads: Option<usize>,
 }
 
 impl Default for Config {
     fn default() -> Self {
         Self {
-            n_threads: None,
-            ignore_case: true,
-            prefer_prefix: false,
-            match_paths: false,
+            prefer_prefix: true,
+            optimize_for_paths: true,
+            n_threads: Some(
+                std::thread::available_parallelism()
+                    .map(std::num::NonZeroUsize::get)
+                    .unwrap_or(4)
+                    .min(8),
+            ),
         }
     }
 }
 
 impl Config {
-    /// Set the number of threads to use.
-    pub fn n_threads(mut self, n_threads: usize) -> Self {
-        self.n_threads = Some(n_threads);
-        self
-    }
-
-    /// Set whether to ignore case.
-    pub fn ignore_case(mut self, ignore_case: bool) -> Self {
-        self.ignore_case = ignore_case;
-        self
+    /// Create a new configuration with the default values.
+    pub fn new() -> Self {
+        Self::default()
     }
 
     /// Set whether to prefer prefix matches.
@@ -52,20 +46,22 @@ impl Config {
     }
 
     /// Set whether to optimize for matching paths.
-    pub fn match_paths(mut self, match_paths: bool) -> Self {
-        self.match_paths = match_paths;
+    pub fn optimize_for_paths(mut self, optimize_for_paths: bool) -> Self {
+        self.optimize_for_paths = optimize_for_paths;
+        self
+    }
+
+    /// Set the number of threads to use for matching.
+    pub fn n_threads(mut self, n_threads: Option<usize>) -> Self {
+        self.n_threads = n_threads;
         self
     }
 }
 
 impl From<&Config> for nucleo::Config {
     fn from(config: &Config) -> Self {
-        let mut matcher_config = nucleo::Config::DEFAULT;
-        matcher_config.ignore_case = config.ignore_case;
-        matcher_config.prefer_prefix = config.prefer_prefix;
-        if config.match_paths {
-            matcher_config = matcher_config.match_paths();
-        }
-        matcher_config
+        let mut nucleo_config = nucleo::Config::DEFAULT;
+        nucleo_config.prefer_prefix = config.prefer_prefix;
+        nucleo_config
     }
 }