before

```
on   ➜ cargo bench --bench main -- --save-baseline before
    Finished `bench` profile [optimized] target(s) in 0.13s
     Running benches/main.rs (target/release/deps/main-8151fda9d09b6257)
results_list            time:   [16.002 µs 16.095 µs 16.190 µs]
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild

Benchmarking draw: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 5.6s, or reduce sample count to 80.
draw                    time:   [2.0829 ms 2.2517 ms 2.4277 ms]
```

after

```
on   ➜ cargo bench --bench main -- --baseline before
    Finished `bench` profile [optimized] target(s) in 0.08s
     Running benches/main.rs (target/release/deps/main-8151fda9d09b6257)
results_list            time:   [8.2873 µs 8.3418 µs 8.3963 µs]
                        change: [-49.407% -48.880% -48.360%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) low mild

Benchmarking draw: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 5.4s, or reduce sample count to 90.
draw                    time:   [1.3345 ms 1.4388 ms 1.5460 ms]
                        change: [-42.872% -36.105% -29.053%] (p = 0.00 < 0.05)
                        Performance has improved.
```