@alexpasmantier 2025-06-18
Questions and nitpicking aside, bravo for that -50% this is great 😁